chore: pull out pure functions from interpreter
